Improve support for `Interval{<:Integer}`
When intervals are components of bigger `struct`s that impose
`eltype`-consistency, it would be nice not to have to force the
`eltype` to be an `AbstractFloat`. It's been possible to construct
`Interval{<:Integer}` for some time, but many methods throw errors.
This reduces the fragility of the package, hopefully without introducing
new numeric errors.