Update docs link in README.md
I'm not sure why the release process isn't producing any docs at `stable` (or, in fact, for any previous versions). 

In the meantime, it seems better to link users to the dev docs rather than a dead link.

FWIW, I had a similar issue in my package some time ago (https://github.com/vtjeng/MIPVerify.jl/pull/116) but it seems to have to do with me using an old version of `Documenter`, which is not the issue you are having here.